+++ c4j-acceptancetest/src/main/java/de/andrena/c4j/acceptancetest/point/ColoredPointSpecContract.java
@@ -14,13 +14,13 @@ public class ColoredPointSpecContract implements ColoredPointSpec {
 	@Override
 	public int getX() {
 		// No contracts identified yet
-		return ignored();
+		return (Integer) ignored();
 	}
 
 	@Override
 	public int getY() {
 		// No contracts identified yet
-		return ignored();
+		return (Integer) ignored();
 	}
 
 	@Override
